selfstorage/requirements.txt
Django==4.2.7
django-environ==0.11.2
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
